move session provider from media to local user data source
