+++ app/src/main/java/com/giraffe/cineverseapp/data/network/SessionProviderImp.kt
@@ -1,0 +1,15 @@
+package com.giraffe.cineverseapp.data.network
+
+import com.giraffe.user.datastore.SessionProvider
+
+class SessionProviderImp: SessionProvider {
+
+    @Volatile
+    private var sessionId: String? = null
+
+    override fun getSessionId(): String? = sessionId
+
+    override fun setSessionId(sessionId: String?) {
+        this.sessionId = sessionId
+    }
+}
