--- app/src/main/java/com/giraffe/cineverseapp/data/util/SessionProviderImpl.kt
@@ -1,15 +1,0 @@
-package com.giraffe.cineverseapp.data.util
-
-import com.giraffe.media.util.SessionProvider
-
-class SessionProviderImpl: SessionProvider {
-
-    @Volatile
-    private var sessionId: String? = null
-
-    override fun getSessionId(): String? = sessionId
-
-    fun setSessionId(sessionId: String?) {
-        this.sessionId = sessionId
-    }
-}
