+++ datasource/local/user/src/main/java/com/giraffe/user/datastore/SessionProvider.kt
@@ -1,0 +1,6 @@
+package com.giraffe.user.datastore
+
+interface SessionProvider {
+    fun getSessionId(): String?
+    fun setSessionId(sessionId: String?)
+}
