--- datasource/remote/media/src/main/java/com/giraffe/media/util/SessionProvider.kt
@@ -1,5 +1,0 @@
-package com.giraffe.media.util
-
-interface SessionProvider {
-    fun getSessionId(): String?
-}
